+++ tasks/add-preview/task.py
@@ -1,25 +1,16 @@
-import boto3
-import json
-import logging
 import numpy
-import requests
+import os
 import shutil
 import tempfile
 
-from boto3utils import s3
-from os import getenv, path as op
 import gdal
+import rasterio
+from boto3utils import s3
 from cirruslib import Catalog, get_task_logger
 from cirruslib.transfer import download_item_assets, upload_item_assets
-
-import os
 from rio_cogeo.cogeo import cog_translate
 from rio_cogeo.profiles import cog_profiles
-import rasterio
 from rasterio.warp import calculate_default_transform, reproject as _reproject, Resampling
-from rasterio.io import MemoryFile
-from urllib.parse import urlparse
-from traceback import format_exc
 
 
 def handler(payload, context={}):
@@ -31,6 +22,7 @@ def handler(payload, context={}):
     outopts = catalog['process'].get('output_options', {})
     assets = config.pop('assets', None)
     thumb = config.pop('thumbnail', False)
+    config.pop('batch')
 
     if assets is None:
         msg = f"add-preview: no asset specified for preview"
@@ -59,12 +51,14 @@ def handler(payload, context={}):
 
             # download asset
             item = download_item_assets(item, path=tmpdir, assets=[asset])
+            filename = item['assets'][asset]['href']
 
             # add preview to item
-            item['assets']['preview'] = create_preview(item, logger, fnout=item['assets'][asset]['href'], **config)
+            item['assets']['preview'] = create_preview(filename, logger, fnout=item['assets'][asset]['href'], **config)
             if thumb:
+                filename = item['assets']['preview']['href']
                 # add thumbnail to item
-                item['assets']['thumbnail'] = create_thumbnail(item, item['assets']['preview']['href'], logger)
+                item['assets']['thumbnail'] = create_thumbnail(filename, item['assets']['preview']['href'], logger)
 
             # put back original href
             item['assets'][asset]['href'] = href
@@ -125,7 +119,7 @@ def calculate_ccc_values(filename, logger, lo=2.0, hi=96.0, bins=1000):
 
 def create_preview(filename, logger, fnout=None, preproj=False, ccc=[2.0, 98.0], exp=None, nodata=0, **kwargs):
     if fnout is None:
-        fnout = op.splitext(filename)[0] + '_preview.tif'
+        fnout = os.path.splitext(filename)[0] + '_preview.tif'
     fntmp = fnout.replace('.tif', '_tmp.tif')
 
     _filename = filename
@@ -155,9 +149,9 @@ def create_preview(filename, logger, fnout=None, preproj=False, ccc=[2.0, 98.0],
         logger.error(f"Unable to create preview {filename}: {err}")
         raise(err)
     finally:
-        if op.exists(fntmp):
+        if os.path.exists(fntmp):
             os.remove(fntmp)
-        if preproj and op.exists(_filename):
+        if preproj and os.path.exists(_filename):
             os.remove(_filename)
 
     return {